back-end/requirements.txt
django
channels
daphne
psycopg2-binary>=2.9
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pillow
requests
pyotp
qrcode
